metariboseq updates (#45)
This PR updates the workflows to avoid duplicate processing of metagenomic assemblies which can be shared across samples. It also breaks up the analysis workflow into separable stages to make it easier to work with samples that have low coverage and hence which can't be fully processed - without these changes the overall analysis scripts will fail with hard to diagnose errors.